Separate WASM API bindings from AS wrapper
