+++ as-lib/Makefile
@@ -1,5 +1,7 @@
 .PHONY: docs
 
+FILES = assembly/index.ts assembly/wrapper.ts assembly/mondeto.ts
+
 DOCS_DIR = docs
 docs:
-	npx typedoc --ignoreCompilerErrors --out $(DOCS_DIR) assembly/index.ts assembly/mondeto.ts
+	npx typedoc --ignoreCompilerErrors --out $(DOCS_DIR) $(FILES)
