+++ as-lib/assembly/index.ts
@@ -5,3 +5,5 @@
 //  https://github.com/AssemblyScript/assemblyscript/pull/594
 //  https://github.com/AssemblyScript/assemblyscript/blob/1fcc374e985a1caca0736126a4e3ca6ca7776eac/tests/packages/packages/a/assembly/index.ts#L1
 export * from "./mondeto"
+// Also export wrappers
+export * from "./wrapper"
